refactor: add stacks-common/src/types/sqlite.rs
